sentry: refactor common RPC type conversions
Sentry is using parts of downloader's sentry_type_casts.hpp
which implements interfaces' types.proto helpers.
Those could be useful for other RPCs too.
