+++ node/silkworm/rpc/interfaces/types.hpp
@@ -16,20 +16,14 @@
 
 #pragma once
 
-#include <memory>
-
+#include <intx/intx.hpp>
 #include <types/types.pb.h>
 
-#include <silkworm/rpc/interfaces/types.hpp>
-
-#include "types.hpp"
+#include <silkworm/types/hash.hpp>
 
 namespace silkworm {
 
-std::unique_ptr<types::H256> to_H256(const intx::uint256& orig);
-std::unique_ptr<types::H256> to_H256(const Hash& orig);
-std::unique_ptr<types::H512> to_H512(const Bytes& orig);
-
-Bytes bytes_from_H512(const types::H512& orig);
+Hash hash_from_H256(const types::H256& orig);
+intx::uint256 uint256_from_H256(const types::H256& orig);
 
 }  // namespace silkworm
